Consolidate streaming tests in the TestClient suite
